Move Node type of out wallet api
